feat(web): wrap useSearchParams usage in Suspense for properties and search pages
Move client-only search param logic into client components and wrap them with Suspense on the server pages to satisfy Next.js CSR bailout requirements.

This resolves build-time errors about missing Suspense boundaries for useSearchParams.
