+++ apps/web/src/app/(auth)/forgot-password/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { useState } from "react";
+import { useState } from "react";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
@@ -8,8 +8,8 @@ import AuthHeader from "@/components/auth/auth-header";
 
 export default function ForgotPasswordPage() {
   const [email, setEmail] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [message, setMessage] = useState("");
+  const [loading] = useState(false);
+  const [message] = useState("");
 
   return (
     <div className="flex items-center justify-center px-4 py-12">
